+++ packages/yourloops/lib/notifications/models/notification-context.model.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, Diabeloop
+ * Copyright (c) 2022-2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -37,7 +37,7 @@ export interface NotificationContext {
   cancel: (notificationId: string, teamId?: string, inviteeEmail?: string) => Promise<void>
   inviteRemoteMonitoring: (teamId: string, userId: string, monitoringEnd: Date, referringDoctor?: string) => Promise<void>
   cancelRemoteMonitoringInvite: (teamId: string, userId: string) => Promise<void>
-  getInvitation: (teamId: string) => Notification
+  getInvitation: (teamId: string, patientEmail: string) => Notification
   refreshReceivedInvitations: () => Promise<void>
   refreshSentInvitations: () => Promise<void>
 }