+++ CHANGELOG.md
@@ -34,6 +34,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2322 Pending patients should not display glycemia specific data
 - YLP-2323 Add new columns from existing data in patient list
 - YLP-2324 Hide certain columns when private practice is selected
+- YLP-2358 As a care team member, I want a new design when listing my pending patients
 
 ### Engineering use
 - YLP-2044 Remove redux settings store
@@ -86,7 +87,10 @@ It is based on Tidepool Blip 1.27.
 - YLP-2307 Incomplete alert when accepting a direct share invitation
 - YLP-2311 Reorganize ITs
 - YLP-2328 Migrate patient-data from jsx to tsx
+- YLP-2364 Fixed English translations
 - YLP-2367 Patient list refacto
+- YLP-2386 Fix navigation crash between care team and patient list
+- YLP-2387 Fix the tab title when displaying the patient list
 
 ### Fixed
 - YLP-2145 Patient cannot see monitoring widgets when he's in a monitoring team and another non-monitoring team